Character Repetition Intervals (CRIs)...
...as an indicator for text similarity.

Where "similarity" is understood in a very narrow sense, actually meaning the same text at different stages of development.
You start writing a text. You save it. You add more text. Save it again. Fix a typo, change sentence order. Save it again. Some part of the text supposedly remains unchanged between the save actions.
And consequently, the distances (intervals) between individual characters remain the same for most part.
This intuition is captured in a matrix of the lengths of intervals between repetitions of a configurable range of characters. The first interval is the number of characters that are skipped before the target character occurs for the first time.
That matrix is reduced to one with a grouping by interval lengths, counting up the lengths in each group.
So, for each symbol you have the number of times the CRI length was 0, 1, 2, etc...
The simplest approach to turning such a matrix into a measure to compare the similarity of two texts is to calculate a diff between the CRI counts of two texts and take it as an absolute distance.

No clue whether this algorithm has any academic name or discourse. If it has, let us know and we'll add the appropriate references.
For the time being it's just a nice playground to learn Kotlin.