A soupçon of evaluation.
How good is the approach? No way we're sitting down evaluating this manually.

But you can download the Wikipedia.
If scores for comparison of different versions of the same article are smaller than scores from comparing the latest versions of distinct articles that should be a good indicator.

Added a little experiment (manual usage) that gives the first impression that this is actually pretty much the case.